:truck: Move or rename resources (e.g.: files, paths, routes).
