+++ src/test/java/org/api/testing/demo/steps/booking/StepsDefinitions.java
@@ -1,10 +1,11 @@
-package org.api.testing.demo.steps;
+package org.api.testing.demo.steps.booking;
 
 import io.cucumber.java.es.Cuando;
 import io.cucumber.java.es.Dado;
 import io.cucumber.java.es.Entonces;
 import net.serenitybdd.screenplay.actors.OnStage;
 import net.serenitybdd.screenplay.rest.abilities.CallAnApi;
+import org.api.testing.demo.exceptions.GenericRuntimeException;
 import org.api.testing.demo.tasks.booking.CreateBooking;
 
 import java.util.List;
@@ -17,24 +18,22 @@ import static org.api.testing.demo.exceptions.AssertionsServices.*;
 import static org.api.testing.demo.questions.common.GetExpectedJsonSchema.theJsonSchemaExpectIs;
 import static org.api.testing.demo.questions.common.GetResponseTime.responseTimeIs;
 import static org.api.testing.demo.questions.common.GetStatusCode.httpResponseStatusCodeIs;
-import static org.api.testing.demo.steps.conf.Actors.CAMILA;
+import static org.api.testing.demo.steps.hooks.Actors.CAMILA;
 import static org.api.testing.demo.utils.common.JsonUtils.parseJsonObject;
 import static org.api.testing.demo.utils.constants.Constants.NAME_SCHEMA_RESOURCE_CREATED;
-import static org.api.testing.demo.utils.constants.Constants.URI;
+import static org.api.testing.demo.utils.constants.Constants.THE_REST_API_BASE_URL;
 import static org.api.testing.demo.utils.enums.HttpStatusCodes.OK;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.lessThanOrEqualTo;
 
-
 public class StepsDefinitions {
-
     public static final String BODY = "body";
 
     @Dado("que la/el cliente desea crear/consultar/actualizar/eliminar una/la reservación de/para su próximo viaje")
     public void preparingAPI() {
         OnStage.theActorCalled(CAMILA.toString());
         theActorInTheSpotlight()
-                .whoCan(CallAnApi.at(CAMILA.recall(URI)));
+                .whoCan(CallAnApi.at(CAMILA.recall(THE_REST_API_BASE_URL)));
     }
 
     @Cuando("el/ella ingresa la siguiente información solicitada para la {} de la reserva")
@@ -63,7 +62,7 @@ public class StepsDefinitions {
             System.out.println("Opción:" + requestOption);
 
         } else {
-            //throw new NotOperationWLFoundException(optionWL);
+            throw new GenericRuntimeException();
         }
     }
