+++ src/test/resources/serenity.conf
@@ -6,31 +6,20 @@ serenity {
   fail.on = java.lang.IllegalAccessError
 }
 
-environment = dev
+# Environments
+environment = restapi
 
 environments {
-  stg {
-    baseurl = "https://restful-booker.herokuapp.com/"
+  restapi {
+    baseUrl = "https://restful-booker.herokuapp.com"
     auth = "auth"
-    getBookingIds = "booking"
-    getBookingId = "booking/:id"
-    getBookingByNames = "booking?firstname={{firstname}}&lastname={{lastname}}"
-    createBooking = "booking"
-    updateBookingId = "booking/:id"
-    partialUpdateBooking = "booking/:id"
-    deleteBooking = "booking/:id"
-    healthCheckApi = "ping"
-  }
-  dev {
-    baseurl = "https://restful-booker.herokuapp.com/"
-    auth = "auth"
-    getBookingIds = "booking"
-    getBookingId = "booking/:id"
-    getBookingByNames = "booking?firstname={{firstname}}&lastname={{lastname}}"
-    createBooking = "booking"
-    updateBookingId = "booking/:id"
-    partialUpdateBooking = "booking/:id"
-    deleteBooking = "booking/:id"
-    healthCheckApi = "ping"
+    getBookingIds = "/booking"
+    getBookingId = "/booking/:id"
+    getBookingByNames = "/booking?firstname={{firstname}}&lastname={{lastname}}"
+    createBooking = "/booking"
+    updateBookingId = "/booking/:id"
+    partialUpdateBooking = "/booking/:id"
+    deleteBooking = "/booking/:id"
+    healthCheckApi = "/ping"
   }
 }
